Fix excerpt
Parse the content as HTML. I noticed that a few excerpts were empty
and I think that's because we're doing a full recovery on parsing
resulting in always getting back at least '<?xml version="1.0"?>'
which meant that the second option (because we used to try XML parsing
before HTML parsing) never got called. So now we're trying just HTML
parsing.

And now we're removing unbreakable spaces and self-made dashes from
the excerpt, too.