Add feed template and delete entry template
Get rid of the distinction between page.html and post.html. This
simplifies some code. It also gets rid of the naming convention: the
first HTML file named on the command line is the output file and the
second one is the HTML template. Both are optional. They default to
index.html and template.html.

Generate the feed using the feed.rss template. This allows people to
generate an Atom feed instead. All they need to do is change the
template. The first RSS/ATOM/XML file is the output file and the
second one is the feed template. Both are optional. They default to
feed.xml and feed.rss.

This means that OPML files can no longer have the .xml extension.