Handle duplicate entries
Atom feeds don't define an order for their entries (RFC 4287, 4.1.2)
so we need to sort the entries ourselves. This is problematic for RSS
feeds that imply a sort order and have an optional pubDate. Adding the
sorting broke some tests. In order to make sorting easier, items
without a pubDate are assigned the start of the epoch (1970-01-01).

In order to eliminate duplicates we have a similar problem. For Atom
entries, the id is mandatory, but for RSS the guid is not. So we pick
the guid, or the link, or the title, or the description (one of the
last three being mandatory).