Use DateTime->compare and UTC for day
I've been using a mix of compare and compare_ignore_floating in the
code. I've switched them all to compare because all the DateTime
objects should have a timezone.

I've also stumbled on as situation like the following:

2020-01-19
<updated>2020-01-19T22:06:51.470-06:00</updated>
2020-01-20
<updated>2020-01-20T03:10:34.277+01:00</updated>
2020-01-19
<updated>2020-01-19T20:05:02.880-05:00</updated>

The correct solution seems to be to convert all the dates to UTC
before computing the YYYY-MM-DD string to use for grouping.