Fix entry content for Atom feeds
Instead of using $xpc->findvalue to find the content (which
automatically removes the tags), use $xpc->findnodes and then use
toString on the result.

Furthermore, in order to remove the surrounding description, content
or summary element, loop through the children and just use toString on
them to assemble the content.