Try harder to find suitable content to show
First, make sure that the description or atom:content nodes we find
aren't empty. If none are found, look for summary, atom:summary, or
itunes:summary elements. This turned out to be necessary because of a
podcast feed that had both empty description elements and the content
in either content:encoded or itunes:summary elements. I figured
itunes:summary was the lesser of two evils, seeing as
https://web.resource.org/rss/1.0/modules/content/ shows dates from
2002 and I have never seen it before.