Revert "Test CDATA section in feed"
This reverts commit 40840cec895b665209793dec74baf5f2d23b90b7.

The result of this test depends on the libxml2 library XML::LibXML was
compiled with. v2.9.14 includes "Fix recovery from invalid HTML start
tags" and this seems to match the observed behaviour: <![CDATA[1<2]]>
results in "1<2" for XML::LibXML built with v2.9.14 but results in "1"
for XML::LibXML built with v2.9.13.

Reference:
https://gitlab.gnome.org/GNOME/libxml2/-/releases/v2.9.14